Started major refactor. Import loop is present
