+++ NetscapeBookmarksFileParser/creator.py
